Remove Celluloid
Celluloid provided us with some nice async abilities and I still think
there are portions of this project that should be async but Celluloid
proved to be more trouble than it was worth. There are no real
requirements for the async behavior and the new code improves
understandability by a lot.